chore: rename footer components
