+++ apps/frontend/src/common/ui/footer/FooterCopyright.vue
