+++ apps/frontend/src/common/ui/footer/FooterSocials.vue
